Get rid of Channel from our firmware and deal with the fallout
Removes the concept of Channel from the firmware.  This change has
large impact on API as it moves the channel definition from being
hardcoded in the firmware to being defined by the RCP application
when it programs the channels.

Still lots of work to do with cleaning up and redicuing size.  This
change is in so that the API get changed to the stable version for
release of MK2.

+++ include/CAN/OBD2.h
@@ -7,7 +7,6 @@
 
 int OBD2_request_PID(unsigned char pid, int *value, size_t timeout);
 void OBD2_set_current_PID_value(size_t index, int value);
-int OBD2_get_current_PID_value(size_t index);
+int OBD2_get_current_PID_value(int index);
 
 #endif /* OBD2_H_ */
-
